Core/Pets: Move Egbert & Pandaren Monk into pet_generic.cpp
(cherry picked from commit 542f5264d379fba9293902ae0c10288d3d660f6d)
